Automated-Price-Comparison-dj_jumia/src2/requirements.txt
appdirs==1.4.4
asgiref==3.3.1
autopep8==1.5.6
beautifulsoup4==4.9.3
black==20.8b1
bs4==0.0.1
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
cssselect==1.1.0
Django==3.1.7
django-bootstrap4==3.0.1
django-chartjs==2.2.1
django-filter==2.4.0
django-model-utils==4.1.1
django-notifications-hq==1.6.0
fake-useragent==0.1.11
idna==2.10
jellyfish==0.8.2
joblib==1.0.1
jsonfield==3.1.0
lxml==4.6.3
mypy-extensions==0.4.3
numpy==1.20.2
pandas==1.2.4
parse==1.19.0
pathspec==0.8.1
Pillow==8.2.0
psycopg2==2.8.6
pycodestyle==2.7.0
pyee==8.1.0
pyppeteer==0.2.5
pyquery==1.4.3
python-dateutil==2.8.1
pytz==2021.1
recordlinkage==0.14
regex==2021.3.17
requests==2.25.1
requests-html==0.10.0
scikit-learn==0.24.1
scipy==1.6.2
six==1.15.0
soupsieve==2.2.1
sqlparse==0.4.1
swapper==1.1.2.post1
threadpoolctl==2.1.0
toml==0.10.2
tqdm==4.59.0
typed-ast==1.4.2
typing-extensions==3.7.4.3
urllib3==1.26.4
w3lib==1.22.0
websockets==8.1
